Video number 16

+++ lib/screens/game/story/Step.dart
@@ -1,7 +1,7 @@
 import 'package:chesterjetpack/screens/game/entities/EntityType.dart';
 
 class Step {
-  final int spawnTime;
+  final double spawnTime;
   final EntityType entityType;
   final List<double> args;
